WIP: Admin functionality PHP page
Utilizing sessions to save the login status of admin, username/pw will be checked against admin account in the database.  This necessitates either another table or another column in the users table.  Recommendation here is a new table in the db rather than reformatting the initial table.  I will implement a new table on the server when I have time available tomorrow.  Info on new table will be posted into the wiki. 

Functionality will allow for admins to pull user information etc.  Will have to revise the initial table though to implement ban status in the long term.  Could utilize the online column that is currently in the users table as it is not currently used for anything, but that creates a clarity issue.

Would have had this up earlier, but it has been an adventure trying to get another system rigged with enough dev tools and getting the web server and php versioning, and a local mysql server to mesh along with getting a VM functional on my laptop.  

Note:  i k